requirements.txt
openai-whisper
flask
openai
python-dotenv
psycopg2-binary
minio
redis
